api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
cryptography>=44.0.0
python-dotenv>=1.0.0
